requirements.txt
torch
diffusers
gradio
pydub
Pillow
matplotlib
numpy
scipy
accelerate
transformers
torchaudio
ipython
tqdm
